CA-382014: blkgetsize: fix return type mismatch
Noticed by an LTO enabled build at link time of a binary using xapi-stdext-unix.

Fix the type and move the prototype declaration to a header file that is included by both C files.
That way a mismatch in the prototype will always result in a compilation error instead of a potential LTO warning.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>